+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-<!--## Unreleased-->
+## Unreleased
+
+* Added `PolymerProject.addPrefetchLinks()` transform.
 <!-- Add new, unreleased changes here. -->
 
 ## [1.3.1] - 2017-05-16